Move KeyVault tests to separate test suite
